Lots of changes here, re-wrote movement controller - it's not working 100%
git-svn-id: http://pocketgnome.googlecode.com/svn/trunk/pocketgnome@208 4a9c585c-cb2d-11dd-a3c7-fda619b15ae3

+++ PlayerDataController.h
@@ -141,7 +141,7 @@ enum ePlayer_RuneTypes {
 - (Position*)deathPosition;
 - (float)directionFacing;
 - (void)setDirectionFacing: (float)direction;
-- (void)setMovementFlags:(UInt32)movementFlags;
+- (void)setMovementFlags:(UInt8)movementFlags;
 - (UInt32)movementFlags;
 - (UInt64)movementFlags64;
 - (void)faceToward: (Position*)position;
